refactor(deshi/inbound): extract InboundHandlerError to errors.ts
PR レビューで「InboundHandlerError は skill-execution-notifications.ts
固有ではなく inbound 全体の共通基盤なので errors.ts に切り出すべき」と
の指摘を受けたため、`src/deshi/inbound/errors.ts` に分離する。

- errors.ts (新規): InboundHandlerError クラスの定義 + 用途 doc コメント
- skill-execution-notifications.ts: クラス定義を削除し ./errors.js から
  import
- host-tools-server.ts: 既存の import path を ./inbound/errors.js に
  張り替え (skill-execution-notifications.js を経由しない直接依存に)

将来 inbound に handler を追加するときも、共通エラー型は errors.ts から
import すれば良い。

local: 全 34 ファイル / 401 テスト pass、tsc 通過

Refs: isbtty/nanoclaw#6
